refactor: extract position sizing logic
