+++ quant_trade/signal/core.py
@@ -141,7 +141,6 @@ from dataclasses import dataclass
 from quant_trade.utils import get_cfg_value, collect_feature_cols, get_feat
 from .utils import (
     softmax,
-    sigmoid,
     smooth_score,
     smooth_series,
     weighted_quantile,
@@ -162,6 +161,7 @@ from .thresholding_dynamic import (
     DynamicThresholdInput,
     compute_dynamic_threshold,
 )
+from .position_sizer import PositionSizerImpl
 
 
 @dataclass
@@ -660,6 +660,9 @@ class RobustSignalGenerator:
         # 风险过滤器实现
         self.risk_filters = RiskFiltersImpl(self)
 
+        # 仓位管理实现
+        self.position_sizer = PositionSizerImpl(self)
+
         # 当多个信号方向过于集中时，用于滤除极端行情（最大同向信号比例阈值）
         # 值由配置 oi_protection.crowding_threshold 控制
 
@@ -873,54 +876,6 @@ class RobustSignalGenerator:
         """更新币种与板块的映射"""
         self.symbol_categories = {k.upper(): v for k, v in mapping.items()}
 
-
-    def compute_tp_sl(
-        self,
-        price,
-        atr,
-        direction,
-        tp_mult: float = 1.5,
-        sl_mult: float = 1.0,
-        *,
-        rise_pred: float | None = None,
-        drawdown_pred: float | None = None,
-        regime: str | None = None,
-    ):
-        """计算止盈止损价格，可根据模型预测值微调"""
-        if direction == 0:
-            return None, None
-        if price is None or not np.isfinite(price):
-            return None, None
-        if price <= 0:
-            return None, None
-        if atr is None or not np.isfinite(atr):
-            return None, None
-        if atr == 0:
-            atr = 0.005 * price
-
-        cfg = getattr(self, "cfg", {})
-        max_sl_pct = get_cfg_value(cfg, "max_stop_loss_pct", 0.05)
-
-        if rise_pred is None:
-            rise_pred = 0.0
-        if drawdown_pred is None:
-            drawdown_pred = 0.0
-
-        if direction == 1:
-            max_sl = price * (1 - max_sl_pct)
-            tp0 = price * (1 + min(rise_pred, 0.10))
-            sl0 = price * (1 + max(drawdown_pred, -max_sl_pct))
-            tp = max(tp0, price * 1.02)
-            sl = min(sl0, max_sl)
-        else:
-            max_sl = price * (1 + max_sl_pct)
-            tp0 = price * (1 - min(rise_pred, 0.10))
-            sl0 = price * (1 - max(drawdown_pred, -max_sl_pct))
-            tp = min(tp0, price * 0.98)
-            sl = max(sl0, max_sl)
-
-        return float(tp), float(sl)
-
     def _base_key(self, k: str) -> str:
         for key in self.delta_params:
             if k.startswith(key):
@@ -1116,153 +1071,6 @@ class RobustSignalGenerator:
         return 0
 
 
-    def compute_exit_multiplier(self, vote: float, prev_vote: float, last_signal: int) -> float:
-        """根据票数变化决定半退出或全平仓位系数"""
-
-        with self._lock:
-            exit_lag = self._exit_lag
-
-        exit_mult = 1.0
-        vote_sign = np.sign(vote)
-        prev_sign = np.sign(prev_vote)
-        if last_signal == 1:
-            if vote_sign == 1 and prev_vote > vote:
-                exit_mult = 0.5
-                exit_lag = 0
-            elif vote_sign <= 0 and prev_sign > 0:
-                exit_lag += 1
-                exit_mult = 0.0 if exit_lag >= self.exit_lag_bars else 0.5
-            else:
-                exit_lag = 0
-        elif last_signal == -1:
-            if vote_sign == -1 and prev_vote < vote:
-                exit_mult = 0.5
-                exit_lag = 0
-            elif vote_sign >= 0 and prev_sign < 0:
-                exit_lag += 1
-                exit_mult = 0.0 if exit_lag >= self.exit_lag_bars else 0.5
-            else:
-                exit_lag = 0
-        else:
-            exit_lag = 0
-
-        with self._lock:
-            self._exit_lag = exit_lag
-
-        return exit_mult
-
-    def _apply_risk_adjustment(self, pos_size: float, risk_score: float) -> float:
-        """根据风险评分调整仓位大小"""
-        risk_factor = math.exp(-self.risk_scale * risk_score)
-        return pos_size * risk_factor
-
-    def _apply_low_volume_penalty(
-        self,
-        pos_size: float,
-        *,
-        regime: str,
-        vol_ratio: float | None,
-        fused_score: float,
-        base_th: float,
-        consensus_all: bool,
-    ) -> tuple[float, bool]:
-        """在低成交量环境下惩罚仓位, 返回是否触发标记"""
-        low_vol_flag = (
-            regime == "range"
-            and vol_ratio is not None
-            and vol_ratio < self.low_vol_ratio
-            and abs(fused_score) < base_th + 0.02
-            and not consensus_all
-        )
-        if low_vol_flag:
-            pos_size *= 0.5
-        return pos_size, low_vol_flag
-
-    def _apply_vol_prediction_adjustment(self, pos_size: float, vol_p: float | None) -> float:
-        """根据预测波动率对仓位进行修正"""
-        if vol_p is not None:
-            pos_size *= max(0.4, 1 - min(0.6, vol_p))
-        return pos_size
-
-    def _adjust_min_pos_vol(self, min_pos: float, atr: float | None, vol_p: float | None) -> float:
-        """根据历史 ATR 或预测波动率调节仓位下限"""
-        vol_ref = 0.0
-        if vol_p is not None and np.isfinite(vol_p):
-            vol_ref = abs(vol_p)
-        elif atr is not None and np.isfinite(atr):
-            vol_ref = abs(atr)
-        return min_pos * (1 + self.min_pos_vol_scale * vol_ref)
-
-    def compute_position_size(
-        self,
-        *,
-        grad_dir: float,
-        base_coeff: float,
-        confidence_factor: float,
-        vol_ratio: float | None,
-        fused_score: float,
-        base_th: float,
-        regime: str,
-        vol_p: float | None,
-        atr: float | None,
-        risk_score: float,
-        crowding_factor: float,
-        cfg_th_sig: dict,
-        direction: int,
-        exit_mult: float,
-        consensus_all: bool = False,
-    ) -> tuple[float, int, float, str | None]:
-        """Calculate final position size and tier.
-
-        当 ``pos_size`` 为零时会返回 ``zero_reason``，其值来源于 :class:`ZeroReason`，
-        便于上层记录仓位被清零的原因。
-        """
-
-        tier = base_coeff * abs(grad_dir)
-        base_size = tier
-
-        zero_reason: str | None = None
-        low_vol_flag = False
-
-        pos_size = base_size * sigmoid(confidence_factor)
-        pos_size = self._apply_risk_adjustment(pos_size, risk_score)
-        pos_size *= exit_mult
-        pos_size = min(pos_size, self.max_position)
-        pos_size *= crowding_factor
-        if direction == 0:
-            pos_size = 0.0
-            # 无趋势方向时不做仓位
-            zero_reason = ZeroReason.NO_DIRECTION.value
-
-        pos_size, low_vol_flag = self._apply_low_volume_penalty(
-            pos_size,
-            regime=regime,
-            vol_ratio=vol_ratio,
-            fused_score=fused_score,
-            base_th=base_th,
-            consensus_all=consensus_all,
-        )
-
-
-        pos_size = self._apply_vol_prediction_adjustment(pos_size, vol_p)
-
-        # ↓ 允许极小仓位，交由风险控制模块再裁剪
-        min_pos = cfg_th_sig.get("min_pos", self.signal_params.min_pos)
-        min_pos = self._adjust_min_pos_vol(min_pos, atr, vol_p)
-        # 当风险评分升高时提升仓位下限，确保高风险环境下更谨慎
-        dynamic_min = min_pos * math.exp(self.risk_scale * risk_score)
-        if self.risk_filters_enabled and pos_size < dynamic_min:
-            # 保留原方向, 仓位限制在动态下限与最大仓位之间
-            pos_size = min(max(pos_size, dynamic_min), self.max_position)
-            zero_reason = ZeroReason.MIN_POS.value
-
-        # 4h 周期 veto 逻辑已停用
-        # if direction == 1 and scores.get("4h", 0) < -self.veto_level:
-        #     direction, pos_size = 0, 0.0
-        # elif direction == -1 and scores.get("4h", 0) > self.veto_level:
-        #     direction, pos_size = 0, 0.0
-
-        return pos_size, direction, tier, zero_reason
 
     # >>>>> 修改：改写 get_ai_score，让它自动从 self.models[...]["features"] 中取“训练时列名”
     def get_ai_score(self, *args, **kwargs):
@@ -2260,130 +2068,6 @@ class RobustSignalGenerator:
             strong_confirm = abs(vote) >= strong_min
             return vote, conf_vote, weak_vote, strong_confirm
 
-    def _determine_direction(
-        self,
-        grad_dir: float,
-        regime: str,
-        fs: dict,
-        st_dir: int,
-        vol_breakout_val: float | None,
-        conf_vote: float,
-        weak_vote: bool,
-        fused_score: float,
-        base_th: float,
-        raw_f1h: dict | None,
-        std_1h: dict,
-        ts,
-        symbol,
-    ) -> int:
-        """Determine final trade direction."""
-
-        if not self.direction_filters_enabled:
-            return 0 if grad_dir == 0 else int(np.sign(grad_dir))
-
-        direction = 0 if grad_dir == 0 else int(np.sign(grad_dir))
-        if weak_vote:
-            direction = 0
-
-        if regime == "range":
-            atr_v = (raw_f1h or std_1h).get("atr_pct_1h")
-            bb_w = (raw_f1h or std_1h).get("bb_width_1h")
-            low_vol = False
-            if atr_v is not None and atr_v < 0.005:
-                low_vol = True
-            if bb_w is not None and bb_w < 0.01:
-                low_vol = True
-            if low_vol:
-                direction = 0
-            elif vol_breakout_val is None or vol_breakout_val <= 0 or conf_vote < 0.15:
-                direction = 0
-
-        if self._cooldown > 0:
-            self._cooldown -= 1
-
-        if self._last_signal != 0 and direction != 0 and direction != self._last_signal:
-            flip_th = max(base_th, self.flip_coeff * abs(self._last_score))
-            if abs(fused_score) < flip_th or self._cooldown > 0:
-                direction = self._last_signal
-            else:
-                self._cooldown = 2
-
-        align_count = 0
-        if direction != 0:
-            for p in ("1h", "4h", "d1"):
-                if np.sign(fs[p]["trend"]) == direction:
-                    align_count += 1
-            if st_dir != 0 and st_dir == direction:
-                align_count += 1
-            min_align = self.min_trend_align if regime == "trend" else max(
-                self.min_trend_align - 1, 0
-            )
-            if align_count < min_align:
-                direction = 0
-
-        return direction
-
-    def _apply_position_filters(
-        self,
-        pos_size: float,
-        direction: int,
-        *,
-        weak_vote: bool,
-        funding_conflicts: int,
-        oi_overheat: bool,
-        risk_score: float,
-        logic_score: float,
-        base_th: float,
-        conflict_filter_triggered: bool,
-        zero_reason: str | None,
-    ) -> tuple[float, int, str | None, list[str]]:
-        """Apply filters to position size and direction."""
-
-        penalties: list[str] = []
-
-        if not self.direction_filters_enabled:
-            return pos_size, direction, zero_reason, penalties
-
-        if weak_vote:
-            if self.filter_penalty_mode:
-                pos_size *= self.penalty_factor
-                penalties.append(ZeroReason.VOTE_PENALTY.value)
-                zero_reason = None
-            else:
-                direction = 0
-                pos_size = 0.0
-                zero_reason = zero_reason or ZeroReason.VOTE_FILTER.value
-
-        if funding_conflicts > self.veto_level:
-            if self.filter_penalty_mode:
-                pos_size *= self.penalty_factor
-                penalties.append(ZeroReason.FUNDING_PENALTY.value)
-                zero_reason = None
-            else:
-                direction = 0
-                pos_size = 0.0
-                zero_reason = zero_reason or ZeroReason.FUNDING_CONFLICT.value
-
-        if oi_overheat:
-            pos_size *= 0.5
-
-        pos_map = base_th * 2.0
-        if risk_score > 1 or logic_score < -0.3:
-            pos_map = min(pos_map, 0.5)
-        pos_size = min(pos_size, pos_map)
-
-        if conflict_filter_triggered:
-            if self.filter_penalty_mode:
-                pos_size *= self.penalty_factor
-                penalties.append(ZeroReason.CONFLICT_PENALTY.value)
-                zero_reason = None
-            else:
-                pos_size = 0.0
-                direction = 0
-                zero_reason = zero_reason or ZeroReason.CONFLICT_FILTER.value
-
-        return pos_size, direction, zero_reason, penalties
-
     def finalize_position(
         self,
         fused_score: float,
@@ -2652,7 +2336,7 @@ class RobustSignalGenerator:
         st_sum = st1 + st4 + stdir
         st_dir = int(np.sign(st_sum)) if st_sum != 0 else 0
 
-        direction = self._determine_direction(
+        direction = self.position_sizer._determine_direction(
             grad_dir,
             regime,
             fs,
@@ -2680,7 +2364,7 @@ class RobustSignalGenerator:
         tier = None
         fused_score = soft_clip(fused_score, k=1.0)
         atr_raw_pre = (raw_f1h or std_1h).get("atr_pct_1h")
-        pos_size, direction, tier, zero_reason = self.compute_position_size(
+        pos_size, direction, tier, zero_reason = self.position_sizer.decide(
             grad_dir=grad_dir,
             base_coeff=base_coeff,
             confidence_factor=confidence_factor,
@@ -2695,14 +2379,14 @@ class RobustSignalGenerator:
             cfg_th_sig=cfg_th_sig,
             direction=direction,
             exit_mult=(
-                self.compute_exit_multiplier(vote, prev_vote, self._last_signal)
+                self.position_sizer.compute_exit_multiplier(vote, prev_vote, self._last_signal)
                 if direction == self._last_signal and self._last_signal != 0
                 else 1.0
             ),
             consensus_all=risk_info.get("consensus_all", False),
         )
 
-        pos_size, direction, zero_reason, penalties = self._apply_position_filters(
+        pos_size, direction, zero_reason, penalties = self.position_sizer._apply_position_filters(
             pos_size,
             direction,
             weak_vote=weak_vote,
@@ -2725,7 +2409,7 @@ class RobustSignalGenerator:
         atr_abs = max(atr_abs, 0.005 * price)
         take_profit = stop_loss = None
         if direction != 0:
-            take_profit, stop_loss = self.compute_tp_sl(
+            take_profit, stop_loss = self.position_sizer.compute_tp_sl(
                 price,
                 atr_abs,
                 direction,
