+++ tests/test_penalty_mode.py
@@ -55,7 +55,7 @@ def test_position_penalty_mode():
     rsg.filter_penalty_mode = True
     rsg.penalty_factor = 0.5
     rsg.veto_level = 0.4
-    pos, direction, zr, p1 = rsg._apply_position_filters(
+    pos, direction, zr, p1 = rsg.position_sizer._apply_position_filters(
         0.2,
         1,
         weak_vote=True,
@@ -72,7 +72,7 @@ def test_position_penalty_mode():
     assert zr is None
     assert p1 == [ZeroReason.VOTE_PENALTY.value]
 
-    pos2, direction2, zr2, p2 = rsg._apply_position_filters(
+    pos2, direction2, zr2, p2 = rsg.position_sizer._apply_position_filters(
         0.2,
         1,
         weak_vote=False,
