+++ tests/test_scoring_layers.py
@@ -6,6 +6,7 @@ from quant_trade.signal.predictor_adapter import PredictorAdapter
 from quant_trade.signal.factor_scorer import FactorScorerImpl
 from quant_trade.signal.fusion_rule import FusionRuleBased
 from quant_trade.signal.risk_filters import RiskFiltersImpl
+from quant_trade.signal.position_sizer import PositionSizerImpl
 
 
 def make_simple_rsg():
@@ -50,6 +51,7 @@ def make_simple_rsg():
     rsg.fuse = rsg.fusion_rule.fuse
     rsg.fuse_multi_cycle = rsg.fusion_rule.fuse
     rsg.risk_filters = RiskFiltersImpl(rsg)
+    rsg.position_sizer = PositionSizerImpl(rsg)
     return rsg
 
 
@@ -60,7 +62,7 @@ def test_layer_scores_product():
     rsg.factor_scorer.score = lambda f,p:{k:0 for k in rsg.base_weights if k!='ai'}
     rsg.combine_score = lambda ai,fs,w=None: ai
     rsg.dynamic_threshold = lambda *a,**k: (0, 0)
-    rsg.compute_tp_sl = lambda *a,**k:(0,0)
+    rsg.position_sizer.compute_tp_sl = lambda *a,**k:(0,0)
     rsg.models={'1h':{'up':None,'down':None},'4h':{'up':None,'down':None},'d1':{'up':None,'down':None}}
 
     feats={'close':100,'atr_pct_1h':0,'adx_1h':0,'funding_rate_1h':0}
