+++ tests/test_signal_generator.py
@@ -9,29 +9,29 @@ from quant_trade.signal.core import compute_dynamic_threshold
 
 def test_compute_tp_sl():
     rsg = make_dummy_rsg()
-    tp, sl = rsg.compute_tp_sl(100, 10, 1)
+    tp, sl = rsg.position_sizer.compute_tp_sl(100, 10, 1)
     assert tp == pytest.approx(102)
     assert sl == pytest.approx(95)
 
-    tp, sl = rsg.compute_tp_sl(100, 10, -1)
+    tp, sl = rsg.position_sizer.compute_tp_sl(100, 10, -1)
     assert tp == pytest.approx(98)
     assert sl == pytest.approx(105)
 
 
 def test_compute_tp_sl_regime():
     rsg = make_dummy_rsg()
-    tp, sl = rsg.compute_tp_sl(100, 10, 1, regime="range")
+    tp, sl = rsg.position_sizer.compute_tp_sl(100, 10, 1, regime="range")
     assert tp == pytest.approx(102)
     assert sl == pytest.approx(95)
 
-    tp, sl = rsg.compute_tp_sl(100, 10, -1, regime="trend")
+    tp, sl = rsg.position_sizer.compute_tp_sl(100, 10, -1, regime="trend")
     assert tp == pytest.approx(98)
     assert sl == pytest.approx(105)
 
 
 def test_compute_tp_sl_min_stop_loss():
     rsg = make_dummy_rsg()
-    tp, sl = rsg.compute_tp_sl(100, 10, 1, sl_mult=0.2)
+    tp, sl = rsg.position_sizer.compute_tp_sl(100, 10, 1, sl_mult=0.2)
     assert tp == pytest.approx(102)
     assert sl == pytest.approx(95)
 
@@ -251,7 +251,7 @@ def test_dynamic_threshold_use_raw_features():
     }
     rsg.combine_score = lambda ai, fs, weights=None: 0
     rsg.dynamic_weight_update = lambda: rsg.base_weights
-    rsg.compute_tp_sl = lambda *a, **k: (None, None)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (None, None)
     rsg.models = {
         '1h': {'up': None, 'down': None},
         '4h': {'up': None, 'down': None},
@@ -451,7 +451,7 @@ def test_generate_signal_with_external_metrics():
         r.combine_score = lambda ai, fs, weights=None: 0.5
         r.dynamic_weight_update = lambda: r.base_weights
         r.dynamic_threshold = lambda *a, **k: (0.0, 0.0)
-        r.compute_tp_sl = lambda *a, **k: (0, 0)
+        r.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
         r.models = {'1h': {'up': None, 'down': None},
                     '4h': {'up': None, 'down': None},
                     'd1': {'up': None, 'down': None}}
@@ -470,7 +470,7 @@ def test_generate_signal_with_external_metrics():
     rsg.combine_score = base.combine_score
     rsg.dynamic_weight_update = lambda: rsg.base_weights
     rsg.dynamic_threshold = base.dynamic_threshold
-    rsg.compute_tp_sl = base.compute_tp_sl
+    rsg.position_sizer.compute_tp_sl = base.position_sizer.compute_tp_sl
     rsg.models = base.models
 
     gm = {'btc_dom_chg': 0.1, 'mcap_growth': 0.1, 'vol_chg': 0.1}
@@ -504,7 +504,7 @@ def test_hot_sector_influence():
     rsg.combine_score = lambda ai, fs, weights=None: 0.5
     rsg.dynamic_weight_update = lambda: rsg.base_weights
     rsg.dynamic_threshold = lambda *a, **k: (0.0, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {'1h': {'up': None, 'down': None},
                   '4h': {'up': None, 'down': None},
                   'd1': {'up': None, 'down': None}}
@@ -546,7 +546,7 @@ def test_eth_dominance_influence():
     rsg.combine_score = lambda ai, fs, weights=None: 0.5
     rsg.dynamic_weight_update = lambda: rsg.base_weights
     rsg.dynamic_threshold = lambda *a, **k: (0.0, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {'1h': {'up': None, 'down': None},
                   '4h': {'up': None, 'down': None},
                   'd1': {'up': None, 'down': None}}
@@ -589,7 +589,7 @@ def test_short_momentum_and_order_book():
     rsg.combine_score = lambda ai, fs, weights=None: 0.5
     rsg.dynamic_weight_update = lambda: rsg.base_weights
     rsg.dynamic_threshold = lambda *a, **k: (0.0, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {'1h': {'up': None, 'down': None},
                   '4h': {'up': None, 'down': None},
                   'd1': {'up': None, 'down': None}}
@@ -634,7 +634,7 @@ def test_ma_cross_logic_amplify():
     rsg.combine_score = lambda ai, fs, weights=None: 0.5
     rsg.dynamic_weight_update = lambda: rsg.base_weights
     rsg.dynamic_threshold = lambda *a, **k: (0.0, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {'1h': {'up': None, 'down': None},
                   '4h': {'up': None, 'down': None},
                   'd1': {'up': None, 'down': None}}
@@ -735,7 +735,7 @@ def test_order_book_momentum_threshold():
     rsg.combine_score = lambda ai, fs, weights=None: 0.5
     rsg.dynamic_weight_update = lambda: rsg.base_weights
     rsg.dynamic_threshold = lambda *a, **k: (0.0, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {
         '1h': {'up': None, 'down': None},
         '4h': {'up': None, 'down': None},
@@ -783,7 +783,7 @@ def test_sentiment_reweight_and_guard():
 
     rsg.factor_scorer.score = fs_func
     rsg.dynamic_threshold = lambda *a, **k: (0.1, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {'1h': {'up': None, 'down': None},
                   '4h': {'up': None, 'down': None},
                   'd1': {'up': None, 'down': None}}
@@ -817,7 +817,7 @@ def test_volume_and_funding_penalties():
 
     rsg.factor_scorer.score = fs_func
     rsg.dynamic_threshold = lambda *a, **k: (0.1, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {'1h': {'up': None, 'down': None},
                   '4h': {'up': None, 'down': None},
                   'd1': {'up': None, 'down': None}}
@@ -860,7 +860,7 @@ def test_momentum_alignment_disables_confirm():
 
     rsg.factor_scorer.score = fs_func
     rsg.dynamic_threshold = lambda *a, **k: (0.1, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {'1h': {'up': None, 'down': None},
                   '4h': {'up': None, 'down': None},
                   'd1': {'up': None, 'down': None}}
@@ -901,7 +901,7 @@ def test_crowding_factor_and_dynamic_threshold():
     rsg.factor_scorer.score = fs_func
     rsg.get_dynamic_oi_threshold = lambda pred_vol=None: 0.6
     rsg.dynamic_threshold = lambda *a, **k: (0.1, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {'1h': {'up': None, 'down': None},
                   '4h': {'up': None, 'down': None},
                   'd1': {'up': None, 'down': None}}
@@ -940,7 +940,7 @@ def test_step_exit_with_order_book_flip():
     rsg.factor_scorer.score = lambda f, p: {k: 0 for k in rsg.base_weights if k != 'ai'}
     rsg.combine_score = lambda ai, fs, weights=None: 0.6
     rsg.dynamic_threshold = lambda *a, **k: (0.1, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {'1h': {'up': None, 'down': None},
                   '4h': {'up': None, 'down': None},
                   'd1': {'up': None, 'down': None}}
@@ -999,7 +999,7 @@ def test_position_size_range_regime():
     }
     rsg.combine_score = lambda ai, fs, weights=None: ai
     rsg.dynamic_threshold = lambda *a, **k: (0.1, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {'1h': {'up': None, 'down': None},
                   '4h': {'up': None, 'down': None},
                   'd1': {'up': None, 'down': None}}
@@ -1042,7 +1042,7 @@ def test_generate_signal_with_cls_model():
     }
     rsg.combine_score = lambda ai, fs, weights=None: ai
     rsg.dynamic_threshold = lambda *a, **k: (0.0, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.ma_cross_logic = lambda *a, **k: 1.0
     rsg.models = {
         '1h': {'cls': None},
@@ -1081,7 +1081,7 @@ def test_conflict_filter_and_pos_size():
                                           'sentiment': 0, 'funding': 0}
     rsg.combine_score = lambda ai, fs, weights=None: ai
     rsg.dynamic_threshold = lambda *a, **k: (0.10, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {
         '1h': {'up': None, 'down': None},
         '4h': {'up': None, 'down': None},
@@ -1114,7 +1114,7 @@ def test_extreme_indicator_scales_down():
     rsg.factor_scorer.score = lambda f, p: {k: 0 for k in rsg.base_weights if k != 'ai'}
     rsg.combine_score = lambda ai, fs, weights=None: ai
     rsg.dynamic_threshold = lambda *a, **k: (0.10, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {
         '1h': {'up': None, 'down': None},
         '4h': {'up': None, 'down': None},
@@ -1149,7 +1149,7 @@ def test_vote_conflict_scales_score():
     rsg.factor_scorer.score = lambda f, p: {k: 0 for k in rsg.base_weights if k != 'ai'}
     rsg.combine_score = lambda ai, fs, weights=None: ai
     rsg.dynamic_threshold = lambda *a, **k: (0.0, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {
         '1h': {'up': None, 'down': None},
         '4h': {'up': None, 'down': None},
@@ -1188,7 +1188,7 @@ def test_confirm_15m_adjusts_score():
     rsg.factor_scorer.score = lambda f, p: {k: 0 for k in rsg.base_weights if k != 'ai'}
     rsg.combine_score = lambda ai, fs, weights=None: 0.6
     rsg.dynamic_threshold = lambda *a, **k: (0.0, 0.0)
-    rsg.compute_tp_sl = lambda *a, **k: (0, 0)
+    rsg.position_sizer.compute_tp_sl = lambda *a, **k: (0, 0)
     rsg.models = {
         '1h': {'up': None, 'down': None},
         '4h': {'up': None, 'down': None},
